Extract FakeRoomListItem and aRustRoomInfo to their own files.
